requirements.txt
requests>=2.31.0
discord.py>=2.3.2
pytz>=2024.1
python-dotenv>=1.0.0
Pillow>=10.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
boto3>=1.28.0